requirements.txt
astroid==1.6.5
click==6.7
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
dlib==19.13.1
Flask==1.0.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mpmath==1.0.0
networkx==2.2
numpy==1.14.5
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
Pillow==5.2.0
pylint==1.9.2
pyparsing==2.2.0
PyQt5==5.11.2
PyQt5-sip==4.19.12
PySide2==0.0.0a1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
scikit-image==0.14.0
scipy==1.1.0
selenium==3.14.0
six==1.11.0
SQLAlchemy==1.2.11
sympy==1.3
toolz==0.9.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
